requirements.txt
python-gitlab~=3.6.0
PyGithub~=1.55